Added m1 with submodels m1.p0, m1.p1.
The interaction previously found between parasite and nestday is not present in this datasat for whatever reason, but there is an interaction between total.eggs and parasite: with parasite = 0 total.eggs is highly significant as shown in submodel m1.p0, but insignificant in submodel m1.p1.

Also made some code order changes in the script.